# ⚡ Critical Performance Improvements & Architecture Cleanup