--- a/src/App.jsx
+++ b/src/App.jsx
@@ -1,17 +1,20 @@
-import React from 'react';
+import React, { lazy, Suspense } from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import { Analytics } from '@vercel/analytics/react';
 import { UnifiedAuthProvider } from './components/auth/UnifiedAuthProvider.jsx';
 import ProtectedRoute from './components/auth/ProtectedRoute.jsx';
 import Layout from './components/layout/Layout.jsx';
-import Dashboard from './pages/Dashboard.jsx';
-import Subscriptions from './pages/Subscriptions.jsx';
-import AnalyticsPage from './pages/Analytics.jsx';
-import Settings from './pages/Settings.jsx';
-import AuthPage from './pages/AuthPage.jsx';
 import ErrorBoundary from './components/ErrorBoundary.jsx';
+import Loading from './components/ui/Loading.jsx';
 import { useUnifiedStoreInit } from './hooks/useUnifiedStoreInit.js';
 import './App.css';
+
+// Lazy load pages for better performance
+const Dashboard = lazy(() => import('./pages/Dashboard.jsx'));
+const Subscriptions = lazy(() => import('./pages/Subscriptions.jsx'));
+const AnalyticsPage = lazy(() => import('./pages/Analytics.jsx'));
+const Settings = lazy(() => import('./pages/Settings.jsx'));
+const AuthPage = lazy(() => import('./pages/AuthPage.jsx'));
 
 function App() {
   // Initialize the unified store with Supabase integration
@@ -26,26 +29,30 @@ function App() {
             v7_relativeSplatPath: true
           }}
         >
-          <Routes>
-            {/* Public route for authentication */}
-            <Route path="/auth" element={<AuthPage />} />
-            
-            {/* Protected routes */}
-            <Route path="/*" element={
-              <ProtectedRoute>
-                <Layout>
-                  <Routes>
-                    <Route path="/" element={<Dashboard />} />
-                    <Route path="/subscriptions" element={<Subscriptions />} />
-                    <Route path="/analytics" element={<AnalyticsPage />} />
-                    <Route path="/settings" element={<Settings />} />
-                    {/* Catch-all route for 404 handling */}
-                    <Route path="*" element={<Navigate to="/" replace />} />
-                  </Routes>
-                </Layout>
-              </ProtectedRoute>
-            } />
-          </Routes>
+          <Suspense fallback={<Loading text="Loading page..." />}>
+            <Routes>
+              {/* Public route for authentication */}
+              <Route path="/auth" element={<AuthPage />} />
+              
+              {/* Protected routes */}
+              <Route path="/*" element={
+                <ProtectedRoute>
+                  <Layout>
+                    <Suspense fallback={<Loading text="Loading content..." />}>
+                      <Routes>
+                        <Route path="/" element={<Dashboard />} />
+                        <Route path="/subscriptions" element={<Subscriptions />} />
+                        <Route path="/analytics" element={<AnalyticsPage />} />
+                        <Route path="/settings" element={<Settings />} />
+                        {/* Catch-all route for 404 handling */}
+                        <Route path="*" element={<Navigate to="/" replace />} />
+                      </Routes>
+                    </Suspense>
+                  </Layout>
+                </ProtectedRoute>
+              } />
+            </Routes>
+          </Suspense>
         </Router>
       </UnifiedAuthProvider>
       <Analytics />