--- a/src/components/layout/Layout.jsx
+++ b/src/components/layout/Layout.jsx
@@ -1,16 +1,21 @@
-import React, { useState } from 'react';
+import React from 'react';
+import { shallow } from 'zustand/shallow';
+import useUnifiedStore from '@/store/unified-store';
 import Header from './Header.jsx';
 import Sidebar from './Sidebar.jsx';
 
 const Layout = ({ children }) => {
-  const [sidebarOpen, setSidebarOpen] = useState(false);
+  const { sidebarOpen, setSidebarOpen } = useUnifiedStore(
+    (state) => ({
+      sidebarOpen: state.ui.sidebarOpen,
+      setSidebarOpen: state.setSidebarOpen,
+    }),
+    shallow
+  );
 
   return (
     <div className="min-h-screen bg-gray-50">
-      <Header 
-        isSidebarOpen={sidebarOpen} 
-        setSidebarOpen={setSidebarOpen} 
-      />
+      <Header />
       <Sidebar 
         isOpen={sidebarOpen} 
         onClose={() => setSidebarOpen(false)}