--- a/src/hooks/useUnifiedStoreInit.js
+++ b/src/hooks/useUnifiedStoreInit.js
@@ -10,9 +10,26 @@ export const useUnifiedStoreInit = () => {
   const { initialize, handleAuthStateChange } = useUnifiedStore();
 
   useEffect(() => {
+    // Clear any potentially stale auth data that might cause refresh token errors
+    const clearStaleAuthData = async () => {
+      try {
+        // Clear any stale Supabase auth data
+        const { error } = await supabase.auth.signOut({ scope: 'local' });
+        if (error && !error.message.includes('User not logged in')) {
+          console.warn('⚠️ useUnifiedStoreInit: Warning clearing stale auth data:', error.message);
+        }
+      } catch (error) {
+        console.warn('⚠️ useUnifiedStoreInit: Error clearing stale auth data:', error.message);
+      }
+    };
+
     // Initialize store on app start
     console.log('🚀 useUnifiedStoreInit: Initializing...');
-    initialize();
+    
+    // Clear stale data first, then initialize
+    clearStaleAuthData().then(() => {
+      initialize();
+    });
 
     // Listen for auth state changes with unified handler
     const { data: { subscription } } = supabase.auth.onAuthStateChange(
@@ -27,7 +44,7 @@ export const useUnifiedStoreInit = () => {
       console.log('🧹 useUnifiedStoreInit: Cleaning up auth listener');
       subscription.unsubscribe();
     };
-  }, [initialize, handleAuthStateChange]);
+  }, []); // Empty dependency array - this should only run once on mount
 };
 
 export default useUnifiedStoreInit;