--- a/src/hooks/useUnifiedStoreInit.js
+++ b/src/hooks/useUnifiedStoreInit.js
@@ -1,50 +1,75 @@
-import { useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import useUnifiedStore from '../store/unified-store.js';
 import { supabase } from '../lib/supabase.js';
+
+// Global flag to track if auth listener is already initialized
+// This prevents multiple listeners when React StrictMode causes double effect runs
+let authListenerInitialized = false;
+let globalAuthSubscription = null;
 
 /**
  * Unified hook to initialize the store and handle authentication state changes
  * This replaces both AuthProvider and useStoreInit to eliminate race conditions
+ * Resilient to React StrictMode double-effect runs
  */
 export const useUnifiedStoreInit = () => {
   const { initialize, handleAuthStateChange } = useUnifiedStore();
+  const hasInitialized = useRef(false);
 
   useEffect(() => {
-    // Clear any potentially stale auth data that might cause refresh token errors
-    const clearStaleAuthData = async () => {
-      try {
-        // Clear any stale Supabase auth data
-        const { error } = await supabase.auth.signOut({ scope: 'local' });
-        if (error && !error.message.includes('User not logged in')) {
-          console.warn('⚠️ useUnifiedStoreInit: Warning clearing stale auth data:', error.message);
-        }
-      } catch (error) {
-        console.warn('⚠️ useUnifiedStoreInit: Error clearing stale auth data:', error.message);
-      }
-    };
+    // Prevent double initialization in React StrictMode
+    if (hasInitialized.current && authListenerInitialized) {
+      console.log('⏭️ useUnifiedStoreInit: Already initialized, skipping...');
+      return;
+    }
 
-    // Initialize store on app start
     console.log('🚀 useUnifiedStoreInit: Initializing...');
     
-    // Clear stale data first, then initialize
-    clearStaleAuthData().then(() => {
+    // Initialize the store only once
+    if (!hasInitialized.current) {
       initialize();
-    });
+      hasInitialized.current = true;
+    }
 
-    // Listen for auth state changes with unified handler
-    const { data: { subscription } } = supabase.auth.onAuthStateChange(
-      (event, session) => {
-        console.log('🔄 useUnifiedStoreInit: Auth state changed:', event, session?.user?.email);
-        handleAuthStateChange(event, session);
-      }
-    );
+    // Setup auth listener only once globally
+    if (!authListenerInitialized) {
+      console.log('🔗 useUnifiedStoreInit: Setting up auth state listener...');
+      authListenerInitialized = true;
+      
+      const { data: { subscription } } = supabase.auth.onAuthStateChange(
+        (event, session) => {
+          console.log('🔄 useUnifiedStoreInit: Auth state changed:', event, session?.user?.email);
+          handleAuthStateChange(event, session);
+        }
+      );
+      
+      globalAuthSubscription = subscription;
+    }
 
-    // Cleanup function
+    // Cleanup function - only unsubscribe when component unmounts (not on StrictMode cleanup)
     return () => {
-      console.log('🧹 useUnifiedStoreInit: Cleaning up auth listener');
-      subscription.unsubscribe();
+      console.log('🧹 useUnifiedStoreInit: Effect cleanup called');
+      // Don't unsubscribe immediately - let it persist for app lifetime
+      // Only unsubscribe on actual app unmount
     };
   }, []); // Empty dependency array - this should only run once on mount
+  
+  // Cleanup global subscription on app unmount (window unload)
+  useEffect(() => {
+    const handleUnload = () => {
+      if (globalAuthSubscription) {
+        console.log('🧹 useUnifiedStoreInit: App unloading, cleaning up auth listener');
+        globalAuthSubscription.unsubscribe();
+        globalAuthSubscription = null;
+        authListenerInitialized = false;
+      }
+    };
+    
+    window.addEventListener('beforeunload', handleUnload);
+    return () => {
+      window.removeEventListener('beforeunload', handleUnload);
+    };
+  }, []);
 };
 
 export default useUnifiedStoreInit;