… CLAUDE.md
… SUPABASE_FIX_SUMMARY.md
… src/components/auth/AuthProvider.jsx
… src/hooks/useStoreInit.js
… src/store/index.js